microservice-AI/requirement.txt
# ============================================
# SMART CAREER AI MICROSERVICE - Requirements
# ============================================

# === FASTAPI CORE ===
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# === DATA MODELS & VALIDATION ===
pydantic>=2.9.0
pydantic-settings>=2.4.0

# === HTTP & REQUESTS ===
requests>=2.32.0
httpx>=0.27.0

# === ENVIRONMENT & CONFIG ===
python-dotenv>=1.0.1

# === FILE UPLOAD ===
python-multipart>=0.0.9

# === AI PROVIDERS ===
openai>=1.10.0
langchain>=0.3.0
google-generativeai>=0.8.0

# === NLP & EMBEDDINGS ===
sentence-transformers>=3.0.0
numpy>=2.1.0
scikit-learn>=1.5.0
nltk>=3.9.0

# === VECTOR DATABASE ===
chromadb>=0.5.0

# === CACHE ===
redis>=5.0.8

# === LOGGING ===
loguru>=0.7.2

# === TESTING ===
pytest>=8.0.0
pytest-asyncio>=0.24.0

# === UTILITIES ===
psutil>=5.9.0
typing-extensions>=4.12.0
anyio>=4.4.0
click>=8.1.7